+++ Assets/Scripts/Data/Decorator/FoodDecorator/FoodStats.cs
@@ -5,9 +5,10 @@ public struct FoodStats
     public Mesh Mesh;
     public Color Color;
     public float Size;
-    public float RecoveryValue;
+    public int RecoveryValue;
 
-    public FoodStats(Mesh mesh, Color color, float size, float recoveryValue)
+
+    public FoodStats(Mesh mesh, Color color, float size, int recoveryValue)
     {
         Mesh = mesh;
         Color = color;
